Fixed a few issues with pathing a module and moved where the utility functions are.
